Minor tweaks. Reverted reload statement order. Included menufeatures in reload.
Previously, I had ordered the reload() statements alphabetically. This broke the ability to generate multiple seeds in a row and has been reverted.

Additionally, menufeatures is now included in the reloads. This is because, while generating multiple seeds worked, the game would crash in the menu because menufeatures.common_added needed to be reset.